[bd-lyhp3,bd-iqv1n,bd-7tv0x] HTTP ergonomics + module updates
- Reporting G1+G4 PASS, v3.0.0 (SageDesert)
- TTP G1 PASS, v4.0.0 (SageDesert)
- Subscriptions 5.2.0: SoC bill_run split + G4 utoipa (SageDesert)
- Integrations 2.3.5: remove unused imports (CopperRiver)
- Production 3.1.0: RoutingStep updated_at + migration (CopperRiver)
- BOM module.toml port fix

+++ modules/subscriptions/src/gated_invoice_creation.rs
@@ -190,7 +190,7 @@ pub async fn create_gated_invoice(
 
     let create_req = CreateInvoiceRequest {
         ar_customer_id,
-        amount_cents: price_minor as i32,
+        amount_cents: price_minor,
         billing_period_end: None,
         billing_period_start: None,
         compliance_codes: None,
